Adding parser and tests
Added parent parser and textToHtmlParser along with passing test bank. More to come as I tidy up some ugly code.

+++ BaconInfoGen/BaconBuilder.Tests/MainWindowTest.cs
@@ -35,7 +35,7 @@ namespace BaconBuilder
         public void TearDown() 
         {
             currentTreeDir = null;
-            Directory.Delete(testPath, true);
+            if (testPath != null) Directory.Delete(testPath, true);
             testPath = "";
         }
         #endregion
